\name{clean.email.corpus}
\alias{clean.email.corpus}
\title{Make an email corpus using the 'tm' and 'tm.plugin.mail' packages}
\usage{
clean.email.corpus(email.corpus)
}
\arguments{
  \item{email.corpus}{an email corpus made via
  make.email.corpus}
}
\description{
'make.email.corpus' makes an email corpus from a flattened
directory of email message files.
}
\examples{
# rdevel.corpus <- clean.email.corpus(rdevel.corpus)
}
\keyword{cleaning}
\keyword{corpus}
\keyword{data}
\keyword{email}

