\name{corpus.from.eml}
\alias{corpus.from.eml}
\title{Make an email corpus from an 'eml' directory}
\usage{
corpus.from.eml(directory)
}
\arguments{
  \item{directory}{absolute path to a directory containing
  email messages, one per file, aka 'eml format'}
}
\description{
'corpus.from.eml' makes an email corpus from a directory
containing email messages
}
\examples{
# email.corpus <- corpus.from.eml('/data/rdevel')
}
\keyword{corpus}
\keyword{email}
\keyword{eml}

