\name{make.email.corpus}
\alias{make.email.corpus}
\title{Make an email corpus using the 'tm' and 'tm.plugin.mail' packages}
\usage{
make.email.corpus(DirSource, Permanent = FALSE, dbName = "corpus.db",
  cores = 1)
}
\description{
'make.email.corpus' makes an email corpus from a flattened
directory of email message files.
}
\examples{
# email.corpus <- sf1.cheat.sheet()
}
\keyword{'text}
\keyword{corpus}
\keyword{email}
\keyword{mining'}

