\name{clean.email.corpus}
\alias{clean.email.corpus}
\title{Clean an email corpus using the 'tm' and 'tm.plugin.webmining' packages}
\usage{
clean.email.corpus(email.corpus)
}
\arguments{
  \item{email.corpus}{an email corpus made via
  make.email.corpus}
}
\description{
'clean.email.corpus' cleans an email corpus made by
'make.email.corpus.
}
\examples{
# rdevel.corpus <- clean.email.corpus(rdevel.corpus)
}
\keyword{cleaning}
\keyword{corpus}
\keyword{data}
\keyword{email}

