\name{corpus.from.mbox}
\alias{corpus.from.mbox}
\title{Make an email corpus from an 'mbox' file}
\usage{
corpus.from.mbox(source.file)
}
\arguments{
  \item{source.file}{an input file of emails in 'mbox'
  format, which may be compressed}
}
\description{
'mbox.to.corpus' makes an email corpus from an 'mbox' file,
which may be compressed!
}
\examples{
# email.corpus <- corpus.from.mbox('2013-December.txt.gz')
}
\keyword{corpus}
\keyword{email}
\keyword{mbox}

